data-export/requirements_export.txt
numpy==1.19.5
pandas==1.1.5
progressbar2==3.53.1